docs: move eslint configuration comments into new eslint descriptions
I always found explaining why it's okay to selectively ignore an eslint
error to be quite clumsy. If you only want to disable a rule for a
single line you'd have to put the explanation above the eslint directive
in order to avoid the directive only disabling the rule for your
comment. This made the docs harder to read, as it was unclear whether
the comment was about the line itself or the eslint directive. Now we've
updated to the latest eslint version we can make use of the more
recently added eslint description fields.