test(eslint): use local ESLint configuration file
Previously, our ESLint unit tests were actually using the ESLint
configuration stored at the root of the Tool Kit monorepo to load their
options. This caused an issue when we bumped the versions of our ESLint
packages at the monorepo root and the typescript-eslint plugin was no
longer compatible with the version of ESLint we were using for testing.

This creates flaky tests so we should instead define a custom, simple
ESLint configuration file to be used just for the tests. I generated one
with ESLint's create script that supports modern JS – using plugins for
TypeScript would be unnecessary for this testing. I've also added the
'root' option to the configuration file so that ESLint doesn't keep
traversing up the file structure looking for other configuration files
to merge with, loading the ESLint configuration at the monorepo in the
process.