fix(vault): bump n-fetch version to remove isomorphic-fetch instance
The older version of n-fetch depended on an old version of n-logger
which used isomorphic-fetch to add node-fetch to the global namespace.
This meant that any package which used the vault library would have
fetch inadvertently added to the global namespace/globalThis.
Specifically this caused an issue with the octokit library for
interfacing with GitHub, as it checks to see if there is a fetch
function in the global namespace then falls back to its version of
node-fetch. This is good cross-compatibility behaviour, but it meant any
API calls would throw an error as the version of node-fetch that
isomorphic-fetch was inserting into the global namespace was older and
incompatible with octokit (as it didn't sufficiently conform to the
spec, in this case as its Headers class was not iterable).

Updating to the latest version of n-fetch fixes the issue, and shouldn't
come with any other breaking changes. I note that the only other
instance of isomorphic-fetch being used transitively is in the
next-router plugin which we should put some time into ripping out to
avoid further subtle bugs like this in the future.