fix(orb): use env_var_name type for defining system-code
Since using change-api@1.0.1, it no longer evaluates our system-code
parameter as the value of the environment variable
$CIRCLE_PROJECT_REPONAME. Instead it sends the string
"CIRCLE_PROJECT_REPONAME" which fails the call to the change api (see
slack thread linked below). According to the CircleCI docs (linked
below), there is a parameter type of `env_var_name` which you provide a
string of the environment variable name. This should pass the value of
the $CIRCLE_PROJECT_REPONAME to the change-api call.

Relevant slack thread:
https://financialtimes.slack.com/archives/C02HH3JAQ5V/p1681485994978449?thread_ts=1681483126.496729&cid=C02HH3JAQ5V

CircleCI doc:
https://circleci.com/docs/reusing-config/#environment-variable-name