build: use standard tsconfig base for Node 16
This sets the `target` and `library` compiler options so that we aren't
polyfilling any ECMAScript constructs that Node 16 supports natively,
and the latest Node 16 library functions are type checked properly. It
will be easier to remember to update this base configuration once we
move to Node 18 than updating the options directly.

It also sets some useful recommended default options, including some
that we were already using which I've now removed as they're redundant.