feat(nodemon): enable nodemon plugin to run non-node scripts
Enable the nodemon plugin to accept a `command` option that allows it to run
non-node scripts via the nodemon `--exec` argument.

A use case might be when an system has multiple processes or workers that we
want to run locally. This option would allow us to combine `nodemon` with
`heroku local` to start all processes at the same time.