refactor(circleci): remove redundant code now that we use isMatch
Previously, we had some additional logic to catch when the Tool Kit orb
needed updating in CircleCI config. This was necessary when we'd only
check for Tool Kit hooks in your config, but now that we check the whole
config, we're checking for the orb version regardless so we don't need
this extra check.