feat(circleci): provide diff for what's missing in CircleCI config
Previously, we’d enforce that your CircleCI configuration was installed
properly by checking if all the Tool Kit hooks were mentioned. This was
quite unreliable as we’d sometimes want other parts of the configuration
updated (such as the orb version, which we ended up carving out a
special case for) but the update was never triggered because the hooks
were still present and accounted for. This was changed as part of the
rearchitecting of the plugin, and we now do a much more thorough check,
ensuring that the configuration is a superset of the expected
configuration we calculate in the plugin. However, the error handling
was largely left as is, meaning it assumes that if the installation
check has failed then its because you were missing Tool Kit hooks.

Now, we use the same diffing code that Jest uses for failed assertions
to pretty-print what fields we were expecting were missing in the user's
config. This fixes the issue with the previous approach where the error
wouldn't print anything if the config was missing fields other than Tool
Kit hooks. It also makes it much easier to see what changes need to be
made to satisfy Tool Kit.