fix(circleci): check that the base configuration matches the user's too
We were mistakenly only checking that the config returned by each hook
was a subset of the user's CircleCI config, but we also want to check
that the base configuration we include regardless of hooks also matches.
This is so that we check that fields like the Tool Kit orb version or
the node executor are correct.