fix(circleci): don't depend on all Node versions for joining jobs
This avoids an issue with generated CircleCI configs where depending on
all the Node version matrix jobs when 'fanning-in' with a job that does
not use a matrix meant that we could not attach the workspace as the
`node_modules` directory in the matrix jobs conflicted. The downside now
is that the fanned-in job will still run even if a matrix job for an
older Node version fails. I don't think CircleCI provides any way to
work around this when testing multiple Node versions concurrently
(there's no way to select which job to attach a workspace from, for
instance).