fix(logger): don't separate every flush to hooked fork by newline
By default, winston will append a newline to every log made, and usually
programs will log to stdout once they reach the end of a line.
Previously we'd assume the latter was always the case, and would log to
winston after the flush, stripping the subprocess's newline so that we
wouldn't print a double newline for every end of line. It's actually
possible to stop winston from adding a newline at the end of a log so
let's do that so we can support subprocesses that flush in the middle of
a line (like Jest's test summary reporter).