fix(logger): don't store labels of filtered logs
We use a global variable to keep track of the label used for the last
log so that we can reduce noise in the terminal and only print a label
describing the context when that context changes. However, we track the
label in the formatting function used to convert the metadata object
into a human-readable string, and it turns out this format function is
called for every log, regardless of its log level. We usually don't
actually print logs with verbose and higher log levels, but seeing as
the format function is still called for them, we're still tracking their
labels. This meant that if, for example, we started printing logs from
nodemon but its first log was a filtered out verbose log then we'd never
actually print out a label identifying the context.