chore: fix CircleCI status badge
The CircleCI status badge has stopped loading; I'm not sure when it stopped loading. We were previously using a URL with an API token in so that the status would show when Tool Kit was a private repository, and we should probably have removed that token sooner once it was public? Doesn't matter now as the token is long expired.