fix(circleci): never append executor name to approval job
This keeps the code consistent across jobs that do and do not use Node
matrices. The waiting-for-approval cannot take Node versions as a matrix
as it is not a job defined in the Tool Kit orb, so we don't append the
executor to its name either. This should happen in jobs that don't use
multiple executors too so we don't get CircleCI errors about referring
to a `waiting-for-approval-node` job that has never been defined.