feat(jest): support jest v28 and v29
release notes:
- https://jestjs.io/docs/28.x/upgrading-to-jest28
- https://jestjs.io/docs/upgrading-to-jest29

all the changes are in Jest itself, not the CLI, which means this is safe to release as a patch. existing users of the Jest plugin will still use the version they have in `package-lock.json`. new users will i believe get v29 automatically installed. that's probably fine.