fix(heroku): support new Heroku apps which now include random identifier
Heroku made a change to the naming of the domain apps run on a couple of
months ago. Previously, the domain would be based on the name of the
app, but now a random 12-character string is appended to make them more
unique I guess. This change only affects newly created apps.

We were assuming the previous format when constructing the URL for gtg
tests. This meant that gtg tests have started to fail when deploying new
apps. The correct domain is actually returned by the same API that we
were previously using to check app names/convert to names from an app
ID, so let's use that. This works for apps using the old domain naming
system as well as the new one.

Co-authored-by: Alex Wilson <alex@alexwilson.tech>