feat(core): run hook checks when running tasks if files have changed
We avoid running checks unconditionally so that we don't have to load
all plugins that export hooks every time we run a command, but we still
want to let the user know if their config files have diverged from what
the hooks have installed. Let's only check the configuration when
running commands if we see the files have changed, storing hashes in
.toolkitstate to compare.

For now, the configuration files to check are hardcoded (CircleCI,
package.json, and .toolkitrc.yml) but in the future we could have hooks
register which files they modify and check those.