fix(npm): run npm version before publishing
npm has started rejecting packages that attempt to publish with a
different version defined in its' package.json. Turns out our plugin
attempted to set the version in package.json but it wasn't saved.
Packages have been publishing with a version number in package.json that
differs to the version we state on npm.

This fix removes the use of the internal npm libraries and replaces it
with using the cli directly.