fix(npm): run npm version with --no-git-tag option
npm will try to set a git tag when running npm version. This is causing issues in CI as CircleCI is not setup with a git config (thankfully). We do not need npm to set a git tag and commit as this job will be running in response to a release tag being created on GitHub.