build: allow sandbox directory to stay without package.json
This package is just a place to test Tool Kit code now and thus doesn't
have a package.json of its own (though one can be added by the tester
and it won't be tracked by git). Our clean up script automatically
removes any package without a package.json script to avoid 'zombie
packages' but we can make an exception for the sandbox.