build: bump tsconfig base version
The only change (apart from the new versioning scheme that mirrors Node
releases) is to the 'module' and 'moduleResolution' fields. The update
changes them from 'commonjs' to 'node16' and 'node' to 'node16'
respectively. This does not affect our type resolution logic at all but
it does change the transpilation of one dynamic import from a require()
call to a native dynamic import, something that Node 16 supports. This
has the benefit of allowing custom Tool Kit plugins to be written in
ESM, as they can still be dynamically imported from our CJS code.