fix(create): don't print default string options with quotes
This gave many users the impression that they should also enter the
override value with quotes as well. I think that I originally used
`JSON.stringify` so that arrays would be printed with square brackets,
but on reflection I also think this is a mistake for the same reason we
don't want to print quotes – arrays shouldn't be overridden with
brackets either.