fix: don't suggest upload-assets-to-s3 in additional plugin prompt
upload-assets-to-s3 is already included by frontend-app, and it's basically pointless for any other kind of repo (unless you definitely know you need it, in which case i think having to install it manually is fine?).

leaving this in (especially having it checked by default if you've selected frontend-app) causes unnecessary duplicate plugins.