feat(circleci)!: define CircleCI configs in .toolkitrc.yml
Following on from similar changes to the PackageJson hook, the CircleCI
hook is now configured via Tool Kit options in .toolkitrc.yml files
instead of modifying shared state in JavaScript functions. This creates
a declarative approach instead of the old functional approach, which
should make configuration much easier to reason with and less abstract,
as well as more flexible, allowing for different kinds of workflows.

With this initial commit I've focused on getting the configuration
generated for next-static to be functionally equivalent to what was
generated previously. Subsequent work will be done to make the config
generated better than what was previously generated, and to support a
more diverse range of project types.