fix(cli): make sure postInstall step isn't run when checks are passing
We've previously seen a few issues where tag filters were being added to
the waiting-for-approval job incorrectly (we do not to wait for approval
on releases.)

This was happening on some older repositories as they had an older
version of Tool Kit installed that still used the postInstall function,
which has since been superseded by install state sharing and is
deprecated. However, we saw this postInstall function was being run with
newer versions of Tool Kit too, even though it's supposed to be skipped
if support for state sharing is detected. postInstall will
unconditionally add the tag filter, which is no longer desirable for new
CircleCI configs.

It turned out that this was only happening if you ran `dotcom-tool-kit
--install` on a repository that already had all its configuration up to
date (i.e., all the `check` methods on hooks were returning `true`).
This was because the code to confirm that the legacy postInstall
function can be skipped was inside a block of code only executed when
`check`'s were returning `false`. I can't think of any reason why we
can't move it out of this conditional block and thus fix Tool Kit's
installation behaviour when running `--install` twice consecutively.

Co-authored-by: Emilio de la Torre <emilio.delatorre@ft.com>
Co-authored-by: Alex Cerda <alexandra.cerda@ft.com>
Co-authored-by: Joel Carr <joel.carr@ft.com>