fix(cypress)!: allow Tool Kit options to override Doppler secrets
We should be able to get the Cypress base URL from either Doppler or the
Tool Kit config, but the setting in your config should override what's
in Doppler as this will allow users to test out a different URL in a
separate branch if needed.

This is a breaking change as people may have been relying on the
previous override behaviour (although only cp-content-pipeline seems to
be doing so at this time).