fix(package-json-hook): fix isInstalled check so it handles split paths
Was previously failing for husky's hook options where we specify a
nested field with lodash's dot syntax for paths (i.e., ('husky.hooks')).
This syntax is ignored if you pass in an array instead of a string so we
need to split the string instead (as we also want to be able to support
field names with full stops in).