build(cli): don't include Tool Kit plugins in devDependencies
These were required for various tests we were running. Whilst we still
reference the plugins when testing whether configs are resolved
correctly, the tests still passed if the dependencies are removed. This
is probably either because they still can be found within the workspace,
or the tests don't actually require the plugins to be present to verify
the semantics are correct.

Removing the dependencies means there can be less churn when updating
plugins, particularly in the next branch where the strict peer
dependency ranges mean that updating cli requires every plugin to also
be updated.