fix(logger): stop stripping newlines from hooked subprocess logging
This partially reverts d1374d14, which I think was a mistake as we
already remove the EOL character normally added by winston, and we skip
appending a newline character to messages if it has a `process`
attribute. So it meant that the only newlines that would be preserved
were ones that happened to not be at the end (or start) of a stream
flush.