fix(logger): don't trim 's' from the beginning and end of a line
Because we were declaring a regex within a template literal we needed to
escape the backslash in the `\s` whitespace meta group. I we don't, the
backslash is parsed as an escape character for the template literal, not
the regex, and is replaced with a literal `s` as `\s` is not a valid
escape sequence. This means the regex would search for an `s` instead of
whitespace. This doesn't affect the unescaped `\n` as the escaped `\n`
and the literal `0x0A` ASCII character are (roughly) equivalent in a
regex.

Instead of fixing the incorrect `\s`, let's remove it entirely and not
trim whitespace characters, just newlines, as I'm not sure we actually
want to strip indentation from hooked console logging.