fix(circleci-deploy): remove unused option from e2e-test-review job
No idea why this was in here. Can't find it used in any job in the orb,
so maybe I added it whilst testing the new config tags and forgot to
remove it later? Or maybe we were planning on using it in some new
custom logic in the job that never came together? Regardless, for now,
it should be fine to just read from the Tool Kit config instead.