feat(commitlint): add a Commitlint plugin
This adds a Commitlint plugin, copied from a mix of Content Pipeline and
various Reliability repos - we're now maintaining this same code in
eight different places:

https://github.com/search?q=org%3AFinancial-Times+Commitlint+extends+Task+language%3AJavaScript&type=code&l=JavaScript

I went for a broad range on the peer dependency because this captures
everything we use across the FT and the command line API we're using
doesn't change between these versions.

I haven't added tests (yet?) because it seemed like quite a lot of setup
to test a relatively small amount of code. I'm happy to give it a go but
probs want direction on what to test. I was originally thinking "Do I
need to spin up a GitHub repo in a temporary folder and try committing
to verify that this fails?". Going by your other tests they all seem
more end-to-end than unit and I couldn't find a lot of mocking.

Co-Authored-By: Kara Brightwell <apaleslimghost@users.noreply.github.com>