fix(cli): correctly check for conflicting command tasks
We used to filter out any conflicting hook tasks for hooks that weren't
being used. However, commands are more flexible and any plugin can
associate their tasks with an arbitrary string as a command tasks so we
no longer have a concept of whether a command is 'in use' or not. This
filter was incorrectly comparing command tasks to hook names (e.g.,
CircleCI) so they would never match and all conflicts would always be
filtered out.

Looks like this was originally missed because we were using the webpack
plugin to test for conflicts but the plugin had been misconfigured for
the Tool Kit major update.