fix(circleci): allow workflows with different jobs to be merged
This was previously too strict and would return a conflict if two
hook installations wanted to modify the same workflow. This was quite
disruptive as most plugins want to modify the same 'tool-kit' workflow
and this shouldn't cause any issues as long as the plugins are affecting
different jobs in the workflow.