fix(cli): gracefully handle hook installation errors
We were previously throwing when we encountered an unknown hook option
or an invalid hook option. This resulted in a cryptic stacktrace without
much for the user to use to fix the issue. Instead, let's wrap the
function in a Validated and use properly formatted error messages to
handle these errors.