build: bump TypeScript config to Node 18 versions
This should have been done as part of us dropping support for Node 16 as
this allows us to reduce the number of polyfills generated. It also
means we can start generating ESM files if we want by updating our
packages to use the `module` type. The change to the `module` setting in
`tsconfig.json` requires us to bump ts-jest to a new minor version that
supports the option.