fix(cli): make sure to load plugins via CJS
We recently bumped our tsconfig base to a newer version that switched
the `module` setting from `commonjs` to `node16`. In most cases this had
no effect on the generated code, as CommonJS code will still be emitted
for files that aren't specified to be ESM modules. However, the setting
also changes dynamic imports within CJS code to be preserved whilst
emitted instead of being transpiled to a (wrapped) call to `require`.

Let's explicitly use require instead of a dynamic import when loading
entrypoints to avoid CJS/ESM incompatibilities. I believe we had
previously opted for a dynamic import in order to get plugins loaded
asynchronously, however this was a misunderstanding as the synchronous
require call was still being emitted. I've kept the function as async
still as I expect we'll want to restore the dynamic import at some point
in the future when we support ESM.