fix: use nullish coalescing instead of zod defaults for circleci hook options with defaults
the `.partial()` on the hook option objects is overriding the `.default()` on their properties.
short of splitting the schema into an incoming options schema that's meant to be merged from
multiple sources, and an outgoing schema for the final hook options, this is the most
straightforward way of doing this.