fix: stop adding CWD to the n-test config path
This line, which appeared in c070cc2, breaks test running. This is
because n-test already prepends `process.cwd()` so it doubles up. This
results in something like:

```
/Users/me/code/project/Users/me/code/project/test/smoke.js
```

Kara suggested replacing it with an error if you're not running n-test
from the current working directory. I copied it loosely from 81ebda7
where we're doing the same for the nodemon plugin.