fix(cli): try to always log ToolKitError details if present
When ToolKitErrors bubble up to the root we want to log the error
details to the user as they usually include the most useful contextual
information for them to act on. However, this wasn't always reliably
happening as the `instanceof` check we used to determine whether an
error was a ToolKitError is unreliable across packages if there are
different versions of `@dotcom-tool-kit/error` installed. Let's also use
a heuristic to guess if an error is a ToolKitError so we always log the
details field if present.