feat: allow option schemas to be loaded from plugins
This will allow custom plugins to define their own option schemas, as
well as decouples plugins from the schemas package, which had the
annoying property of causing a cascade of plugins to bump whenever a
schema was updated for a single plugin.

Now we load task and hook options with the same entrypoint as the base
classes are loaded from, looking for an exported `schema` object which
is expected to be a Zod schema. Plugin options are defined in a separate
entrypoint which is specified by the new `optionsSchema` setting.

In order to avoid any breaking changes we'll still fall back to schemas
exported by the `schema` package if a schema isn't found in the plugin
so that bumping the `cli` won't require bumping all plugins too. Note
that the `schemas` package will no longer be maintained and any future
schema updates will be done in the plugins themselves.