backstage(schemas): delete obsolete package
Now that schemas are colocated with plugins, the schema package is no
longer used. Let's delete it so that release-please doesn't
inadvertently bump new releases, and to ensure that people don't
accidentally add schema changes here rather than within plugins.

`core` will still depend on it for now for backwards compatibility, but
it will be use the latest version in the registry rather than from the
workspace.