backstage(create): stop building create package for now
Will need rework now that the schemas package has been removed but it's
been unmaintained since the last major Tool Kit release so not worth
putting in the effort now. But think it could still be valuable in the
future so equally not inclined to delete the package just yet.