test(cli): fix failing tests with an alternative to file path mocks
Now that we resolve the options schemas during the loadConfig function
we can no longer mock out the logic that sets the root paths to use
relative paths, as that causes the resolution to fail. I've hacked in an
alternative that serialises the objects first then does a simple regex
replace.