feat: store pushed Docker images in state
We will need the built image names and tags in order to deploy via Hako.
This involved a bit of refactoring of the tag generation so that we
weren't duplicating stuff. We're using the CI build number for deploys
for now but this can change with no issue later if we change our minds.