fix(orb): automatically use latest patch version of cloudsmith-oidc orb
Code Management just pushed out a hotfix for the orb to account for an
API change with Cloudsmith. Let's not specify the patch version when
specifying their orb so that CircleCI will always use the latest patch
(of the pinned minor version.)

Code Management suggested to do this themselves so we can assume they'll
respect semver guidelines.