feat: add a new AWS plugin
This plugin is used to assume an STS role and provide AWS credentials
for other tasks. It stores AWS credentials in the CI state so they can
be picked up later.

This is required for us to be able to assume different roles across
different accounts with Hako, we'll be deploying to separate test and
prod AWS accounts in future.

Later it makes sense to rework the `serverless` and
`upload-assets-to-s3` plugins to use this. I'd rather not actually do
this work until we've tested this with Hako.

Release-As: 0.1.0
Co-Authored-By: Alex Muller <alexmuller@users.noreply.github.com>