fix(circleci): use correct Tool Kit workspace commands
The persistence job step that is installed into custom jobs was calling
CircleCI's `persist_to_workspace` command directly rather than the Tool
Kit orb abstraction, and the attachment job step was doing the same
except with the kebab casing of the Tool Kit command's name which meant
it would create an invalid config file.

This hasn't caused any problems as of yet as the attachment command is
only installed for custom jobs which aren't particularly used currently.