build: fix reference to old field name in generate-docs script
Based on a code review suggestion, I updated the name of the field
returned from importEntryPoint from entryPoint to baseClass as the
previous name was overloading a parameter name that served a different
purpose. I checked for build errors to find places where the name needed
to be updated, but of course the generate-docs script is written in
JavaScript and not checked by TypeScript, so I missed that the change
needed to be reflected in the script too. This is now fixed.