feat!: adopt standard labels for container images
This removes a lot of the tags we've been adding to images, reducing it
down to only the tag we use for deployment. It replaces many of these
tags with the labels outlined in the Standard Labels for FT Container
Images proposal.

See-also: https://docs.google.com/document/d/1USDbK4ptBe4VRW_3Ujv0RCgrD9JOhMHT1JjImQGvmRQ/edit