feat: support review apps
This adds in an option to run hako deploy with an "ephemeral" app, which
is a similar concept to Heroku review apps. To make this work we need to
pass Hako config into the Docker command as STDIN.

This also updates the two wrapper plugins to introduce a `deploy:review`
command.

This was all tested locally in cp-docker-migration-canary. It works
aside from an issue with name length, we're talking to Cloud Enablement
about this.

Co-Authored-By: Alex Muller <alexmuller@users.noreply.github.com>