feat(serverless): widen dependency range for serverless-offline
The range now includes v13. This version drops support for Node 16
(which we've also dropped support for) and no longer will try to install
node-fetch if fetch is not present.

This is important for us as it uses a top-level await to import
`node-fetch` which causes issues when running with Node 22. Serverless
will typically try to require a module then fall back to an ESM import
if that fails, but with Node 22 require'ing a module that has any top
level awaits will cause an exception to be thrown instead.

Adding v13 to the range of permissible versions will allow consumers to
update their version of serverless-offline in order to avoid this issue
as part of a Node 22 migration. Note that the later v14 is not included
in the range as that require Serverless v4 which we do not support.