fix: make review app URLs predictable
We now generate a hash based on the branch name and use this as a suffix
on review app names. This means a couple of things:

  1. Review app names are predictable
  2. Only one review app will be created per branch, with future commits
     to that branch updating the existing review app instead of
     deploying another one

This is good for a couple of reasons:

We can now hand the review app URL off to future jobs. This is important
for the n-test and cyprus plugins which expect a URL in the `review` or
`staging` state.

We will stop creating a new review app for every commit, we're already
seeing way too many review apps get created which is a problem while we
don't have automatic deletion of old review apps.

Changes to a PR will deploy faster because most of the review app
CloudFormation stack will have been created already - we'll be doing
updates instead of fully creating a new stack on each commit.

See-also: https://financialtimes.atlassian.net/browse/CPREL-1300