fix(hako): use correct load balancer URL based on environment
Previously we were assuming that the load balancer URL we should point
to for tests was the name of the environment + 'ftweb.tech'. This isn't
actually the case – the region prefixes the account name in the URL, as
opposed to being suffixed in the environment name. I've introduced a map
to convert from an environment to a domain. Currently, all domain names
follow the same pattern so we could generate this programmatically but
let's not assume that future domains will follow this pattern too.