feat: add a node-test plugin
This adds a node-test plugin using `node:test#run()`. This plugin only
works in Node.js 20 and 22, with no support for Node.js 18. It's
possible to add support for Node.js 18.17.0+ however it complicates the
plugin a lot - I don't think it's worth doing seeing as 18 goes
end-of-life in April.

The options for this task go against some of the Tool Kit design
principals, mostly the fact that we rely on task options instead of a
central config file. This is because `node --test` is configured
completely via CLI or JS options and there's no config file.

We also make a compromise to support both Node.js 20 _and_ 22. Node.js
22 supports glob-patterns out of the box however Node.js 20 does not.
This means we need to glob for files ourselves. I'm using the same
patterns as Node.js 22's defaults so we can just remove these once we
drop support for Node.js 20.

Co-Authored-By: Ivo Murrell <ivomurrell@users.noreply.github.com>