fix(node-test): don't close stdout once tests are complete
Closing stdout will result in subsequent printing to the console causing
the Tool Kit process to crash with an EPIPE error. This is not
guaranteed to happen as if NodeTest is the only task running in a
command then any subsequent logging tends to only be to stderr. We do
see errors if other tasks run after NodeTest in the command, however.