fix(jest): add support for guessing thread count of cgroupv2 executors
CircleCI recently updated the version of cgroup running on their Docker
executors. This is the logic that allocates a given subset of CPU/memory
resources to our workloads on shared machines. The previous method of
guessing how many virtual CPUs were allocated to us no longer works so
an additional method for cgroupv2 executors is now also attempted.