build: drop type-fest version override
This was originally put in place to work around some TypeScript compiler
errors when importing a PartialDeep type from the circleci plugin. Since
Tool Kit v4, hook options are defined in config rather than in code so
this concern is no longer relevant. Additionally, it was causing
Dependabot to fail as it couldn't handle the override.