fix(logger): explicitly coerce formatting messages into strings
This was highlighted by a winston update that switched the message field
type from any to unknown. We were happily assuming the message was
always actually a string previously but we should be a bit more careful
and explicitly coerce it into one so that calling the prototype methods
actually work as we expect.