feat(cli): add new --validate CLI option
This argument will make Tool Kit eagerly load all configuration, hooks,
plugins, and tasks and make sure everything is specified as it should
be. Since Tool Kit v4, we avoid loading as much code as possible to
optimise startup time, but this lazy approach means that it's difficult
for users to see if there are issues with their configuration until they
actually run a given task, which could only occur late into a CI
workflow. This command sacrifices performance to enable users to
validate their configuration in one-off scenarios.