fix(npm): use correct import for tar module
It needs to be a star import instead of a default import. The tar
package is a TypeScript project which doesn't have a default export.
Because it's TypeScript it will set __esModule to true to tell other
code generated by TypeScript that it doesn't need to wrap the whole
package in a { default: x } object. However, for some reason, TypeScript
doesn't then warn us that there isn't a default object and happily let's
us incorrectly import and call the module as if there is. Not sure why.