feat(containerised-app)!: allow hako environments to be configured
This allows flexibility for when apps want to deploy to custom
environments.

We drop the old `multiregion` option as part of this change as it can
now be better expressed by setting the `hakoProductionEnvironments`
option, rather than carving out a special case for the ft-com
environments. We could have preserved backwards-compatibility with this
change – or even continued supporting the option – but it feels better
to rip the bandage off now whilst not many apps have it set and to
encourage one way of defining multiple environments.